Fix bug when saving "no symptom"
The usage count in DB for symptoms is incremented after during every new symptom event. But the "no symptom" had not DB entry, so the usage update fails.

Quick fix was to put a guard aound the usage update. But a better fix would be to add the "no symptom" as a row in DB